+++ public/js/components/ManagementPanel.js
@@ -1,8 +1,8 @@
-var React = require('react');
-var rB = require('react-bootstrap');
-var Select = require('react-select').default;
-var cE = React.createElement;
-var AppActions = require('../actions/AppActions');
+const React = require('react');
+const rB = require('react-bootstrap');
+const Select = require('react-select').default;
+const cE = React.createElement;
+const AppActions = require('../actions/AppActions');
 
 class ManagementPanel extends React.Component {
     constructor(props) {
@@ -14,10 +14,45 @@ class ManagementPanel extends React.Component {
         this.doTransfer = this.doTransfer.bind(this);
         this.doAccept = this.doAccept.bind(this);
         this.doQuery = this.doQuery.bind(this);
+
+        this.handleAppNameCost = this.handleAppNameCost.bind(this);
+        this.handleCost = this.handleCost.bind(this);
+        this.doChangeCost = this.doChangeCost.bind(this);
+        this.doUpdateUsage = this.doUpdateUsage.bind(this);
+    }
+
+    handleAppNameCost(ev) {
+        AppActions.setLocalState(this.props.ctx, {
+            appNameCost: ev.target.value
+        });
+    }
+
+    handleCost(ev) {
+        let cost = parseFloat(ev.target.value);
+        cost = (isNaN(cost) ? ev.target.value : cost);
+        AppActions.setLocalState(this.props.ctx, {cost});
+    }
+
+    doChangeCost(ev) {
+        if (this.props.appNameCost &&
+            (typeof this.props.appNameCost === 'string') &&
+            (typeof this.props.cost === 'number')) {
+            AppActions.updateApp(this.props.ctx, this.props.appNameCost,
+                                 this.props.cost);
+        } else {
+            console.log('Error: Cannot change cost, missing inputs ' +
+                        JSON.stringify(this.props));
+            AppActions.setError(this.props.ctx,
+                                new Error('Cannot change cost, bad inputs'));
+        }
+    }
+
+    doUpdateUsage(ev) {
+        AppActions.computeAppUsage(this.props.ctx);
     }
 
     handleDeltaUnits(ev) {
-        var deltaUnits = parseFloat(ev.target.value);
+        let deltaUnits = parseFloat(ev.target.value);
         deltaUnits = (isNaN(deltaUnits) ? ev.target.value : deltaUnits);
         AppActions.setLocalState(this.props.ctx, {
             deltaUnits: deltaUnits
@@ -55,7 +90,7 @@ class ManagementPanel extends React.Component {
     }
 
     handleSelectChange(user) {
-        var username = user.value;
+        const username = user.value;
         if (username && (typeof username === 'string')) {
             AppActions.changeUsername(this.props.ctx, username);
         } else {
@@ -67,7 +102,7 @@ class ManagementPanel extends React.Component {
     }
 
     render() {
-        var fields = (this.props.privileged ? [
+        const fields = (this.props.privileged ? [
             cE(rB.Col, { xs: 12, sm: 8, key: 2355},
                cE(rB.FormGroup, {
                    controlId: 'deltaUnitsId'
@@ -89,7 +124,51 @@ class ManagementPanel extends React.Component {
               )
         ] : []);
 
-        var allUsers = this.props.allUsers || [];
+        const fields2 = (this.props.privileged ? [
+            cE(rB.Col, { xs: 8, sm: 4, key: 1355},
+               cE(rB.FormGroup, {
+                   controlId: 'appNameId'
+               },
+                  cE(rB.ControlLabel, null, 'App Name'),
+                  cE(rB.FormControl, {
+                      type: 'text',
+                      value: this.props.appNameCost,
+                      placeholder: 'user-myapp',
+                      onChange: this.handleAppNameCost
+                  })
+                 )
+              ),
+            cE(rB.Col, {xs: 8, sm: 4, key: 1355},
+               cE(rB.FormGroup, {
+                   controlId: 'costId'
+               },
+                  cE(rB.ControlLabel, null, 'Cost'),
+                  cE(rB.FormControl, {
+                      type: 'text',
+                      value: this.props.cost,
+                      placeholder: 'Days per unit',
+                      onChange: this.handleCost
+                  })
+                 )
+              ),
+            cE(rB.Col, {xs: 4, sm: 4, key: 1111},
+               cE(rB.Button, {
+                   onClick: this.doChangeCost,
+                   bsStyle: 'danger'
+               }, 'Change')
+              )
+        ] : []);
+
+        const fields3 = (this.props.privileged ? [
+            cE(rB.Col, {xs: 6, sm: 6, key: 81111},
+               cE(rB.Button, {
+                   onClick: this.doUpdateUsage,
+                   bsStyle: 'danger'
+               }, 'Update App Usage')
+              )
+        ] : []);
+
+        let allUsers = this.props.allUsers || [];
         allUsers = allUsers.map(x => ({value: x, label: x}));
 
         return cE(rB.Grid, {fluid: true},
@@ -141,7 +220,9 @@ class ManagementPanel extends React.Component {
                           )
                        )
                     ),
-                  cE(rB.Row, {className: 'row-center-align'}, fields)
+                  cE(rB.Row, {className: 'row-center-align'}, fields),
+                  cE(rB.Row, {className: 'row-center-align'}, fields2),
+                  cE(rB.Row, {className: 'row-center-align'}, fields3)
                  );
     }
 };