+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes for the `Caf.js` project.
 
 ## 0.4.1
 - Add app resource consumption graphs
+- Track app usage
 
 ## 0.4.0
  - Add new copyright to simplify third-party contributions